Added a better anti-stall method and added better gc collect code
new setting: ApplyAntiFloodForVMUsage, which is true by default and
which will enable a thread sleep so the network layer on the target
server isn't flooded with a lot of queries when executing individual
fetches. VMs suffer from flooding so if you're using a VM, leave it on.
Executing on a real server/locally can have the setting set to false so
benchmarks are completed faster as the sleep is removed.
Thanks to Jonny Bekkum for the gc collect update